Added friendly id's for following entities: - Representative (slug is external_id) - Committee (external_id) - Category (unique name) - District (unique name) - Field (unique name) - Issue (external_id) - Party (external_id) - Vote (external_id)
Friendly id's are stable in the sense that new slugs are not generated when modifying an existing entity, even if the change affects the field the slug was built on.

Run rake db:migrate to update db schema. Slugs are not filled at migration; to fix this, either save entities to create slugs for them ( Representative.fetch_all(&:save) ), or re-import data.