Some clean up of JS sorting on representative#index.
* Moved code into a RepresentativeSorter object
* The previous code would run on all pages, since most stuff
happened in a 'ready' callback in representatives.js. This has
been moved to a content_for block in the representatives_list partial,
which makes sure it only runs on the relevant page.
* Made sort controls into a select list, rendered directly in the erb template instead
of dynamically added by JS.

I did consider just moving the sorting server-side, but we can
always fix that later.