Upgrade chosen. Use chosen.css.
This new version shows disbaled options in a dropdown as greyed out, for /questions/new in #573.
Using chosen.css instead of bootstrap-chosen.less which did not work with this new version of chosen. We may want to change back if he ever upgrades it to support this latest version, but i think this looks ok now too.